Move out server-only types from common
