+++ src/admin/mod.rs
@@ -1,2 +1,3 @@
+pub mod entry;
 pub mod registry;
 pub mod server;
